% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssgac_manhattan.R
\name{ssgac_manhattan}
\alias{ssgac_manhattan}
\title{Create A Manhattan Plot}
\usage{
ssgac_manhattan(sumstat_data_path, clump_data_path, output_filepath, rsID, Chr,
  P.value, position, clump_rsID, clump_Chr, clump_P.value, clump_position,
  delimiter, title)
}
\arguments{
\item{sumstat_data_path}{Summary statistics data path}

\item{clump_data_path}{Plink clumped data path}

\item{output_filepath}{Where you will output the Manhattan Plot}

\item{rsID}{The title of your SNP ID data column in GWAS data (include quotation marks)}

\item{Chr}{The title of your Chromosome data column in GWAS data (include quotation marks)}

\item{P.value}{The title of your p-value data column in GWAS data (include quotation marks)}

\item{position}{The title of your base pair position data column in GWAS data (include quotation marks)}

\item{clump_rsID}{The title of your SNP ID data column in clumped data (include quotation marks)}

\item{clump_Chr}{The title of your Chromosome data column in clumped data (include quotation marks)}

\item{clump_P.value}{The title of your p-value data column in clumped data (include quotation marks)}

\item{clump_position}{The title of your base pair position data column in clumped data (include quotation marks)}

\item{delimiter}{The way your data is separated (often with "/t" or ",")}

\item{title}{What you would like the title to be}
}
\description{
This function intakes GWAS summary statsitcs and produces a Manhattan Plot.
}
\examples{
ssgac_manhattan(sumstat_data_path = "path/sumstata.txt" , 
output_filepath = "outpath", 
rsID="snpid", Chr="chr", 
P.value = "pval", position="bppos",
clump_rsID="snpid", clump_Chr="chr", 
clump_P.value = "P", clump_position="pos",
delimiter = "/t", 
title="ssgac_manhattanplot"))


}
