% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssgac_qq.R
\name{ssgac_qq}
\alias{ssgac_qq}
\title{Create A QQ Plot}
\usage{
ssgac_qq(sumstat_data_path, output_filepath, P.value, delimiter, title)
}
\arguments{
\item{sumstat_data_path}{Summary statistics data path}

\item{output_filepath}{Where you will output the QQ Plot}

\item{P.value}{The title of your p-value data column (include quotation marks)}

\item{delimiter}{The way your data is separated (often with "/t" or ",")}

\item{title}{What you would like the title to be}
}
\description{
This function intakes GWAS summary statsitcs and produces a QQ Plot.
}
\examples{
ssgac_qq(sumstat_data_path = "path/sumstata.txt" , 
output_filepath = "outpath", 
P.value = "pval", delimiter = "/t", 
title="ssgac_qqplot"))

}
